% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAssign.R
\name{doAssign}
\alias{doAssign}
\title{Changing a binding in a specific environment}
\usage{
doAssign(x, value, envir)
}
\arguments{
\item{x}{Variable name given as a \code{character} string}

\item{value}{Value to be assigned to \code{x}}

\item{envir}{Environment in which the binding should be changed}
}
\value{
\code{invisible(NULL)}
}
\description{
Changes a binding in a specific environment
}
\details{
First unlocks the binding and then performs the assignment
in the specified environment
}
