% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAnywhere.R
\name{findAnywhere}
\alias{findAnywhere}
\title{Search Bindings in any Environment}
\usage{
findAnywhere(x)
}
\arguments{
\item{x}{Variable name given as a \code{character} string}
}
\value{
\code{list} of \code{environment}s where a binding
with the specified name has been located
}
\description{
Searches bindings in "any" environment
}
\details{
Searches along the search list and loaded namespaces.
"Along" refers to the fact that enclosing environments are searched
as well. I.e. for a package environment also the package's namespace
environment and a potentially available imports environment are searched
as well.
}
