% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godmode.off.R
\name{godmode.off}
\alias{godmode.off}
\title{Disable god-mode}
\usage{
godmode.off()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Disables god-mode by attempting to re-assign the original versions of praise and encourage.
}
\examples{
godmode.off()
}
