% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godmode.on.R
\name{godmode.on}
\alias{godmode.on}
\title{Enable god-mode}
\usage{
godmode.on(pv = FALSE)
}
\arguments{
\item{pv}{\code{logical} that allows switching to proverbs if set to \code{TRUE}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Enables god-mode by overriding praise and encourage.
}
\examples{
godmode.on()
}
