% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proverbs_a}
\alias{proverbs_a}
\title{English success-related proverbs}
\format{
A character vector, where each element is a different proverb.
}
\source{
\url{https://en.wikiquote.org/wiki/English_proverbs_(alphabetically_by_proverb)}
}
\usage{
proverbs_a
}
\description{
A dataset containing a subset of English proverbs that can be fitting in situations of success,
victory or the like.
}
\keyword{datasets}
