% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proverbs_b}
\alias{proverbs_b}
\title{English non-success-related proverbs}
\format{
A character vector, where each element is a different proverb.
}
\source{
\url{https://en.wikiquote.org/wiki/English_proverbs_(alphabetically_by_proverb)}
}
\usage{
proverbs_b
}
\description{
A dataset containing a subset of English proverbs that can be fitting in situations of near-success,
perseverance, success in the face of adversity, minor failure or the like. Basically whatever you may want to hear
after successfully triggering a test-case failure.
}
\keyword{datasets}
