% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochModel.R
\name{stochModel}
\alias{stochModel}
\title{Stochastic mini-model}
\usage{
stochModel(simSeed = 123)
}
\arguments{
\item{simSeed}{\code{integer} seed. Used to allow simulation of identical results.}
}
\value{
Maximum value contribution statistics for a set of randomly produced variates.
}
\description{
Example function to illustrate godmode usage in testing.
}
\examples{
ret <- stochModel()
}
